refactor: common.py を alias.py と util.py に分割
