+++ src/redi/cli/attachment_command.py
@@ -6,7 +6,7 @@ from redi.api.attachment import (
     read_attachment,
     update_attachment,
 )
-from redi.cli.common import resolve_alias
+from redi.cli.alias import resolve_alias
 from redi.cli.prompt_util import confirm_delete
 from redi.i18n import messages
