+++ src/redi/cli/group_command.py
@@ -1,6 +1,6 @@
 import argparse
 
-from redi.cli.common import resolve_alias
+from redi.cli.alias import resolve_alias
 from redi.cli.prompt_util import confirm_delete
 from redi.i18n import messages
 from redi.api.group import (
