+++ src/redi/cli/util.py
@@ -5,22 +5,6 @@ import tempfile
 from redi.config import editor
 
 
-SUBCOMMAND_ALIASES: dict[str, str] = {
-    "v": "view",
-    "c": "create",
-    "u": "update",
-    "co": "comment",
-    "d": "delete",
-    "l": "list",
-}
-
-
-def resolve_alias(command: str | None) -> str | None:
-    if command is None:
-        return None
-    return SUBCOMMAND_ALIASES.get(command, command)
-
-
 def open_editor(initial_text: str = "") -> str:
     with tempfile.NamedTemporaryFile(suffix=".md", mode="w+", delete=False) as f:
         if initial_text:
